refactor: make package dto
