+++ src/main/java/com/landvibe/landlog/domain/Member.java
@@ -7,13 +7,13 @@ public class Member {
     private String email;
     private String password;
 
-    public Member() {
-    }
-
     public Member(String name, String email, String password) {
         this.name = name;
         this.email = email;
         this.password = password;
+    }
+
+    public Member() {
     }
 
     public Long getId() {
