+++ src/main/java/com/landvibe/landlog/service/MemberService.java
@@ -1,13 +1,11 @@
 package com.landvibe.landlog.service;
 
-import com.landvibe.landlog.ErrorMessage;
-import com.landvibe.landlog.controller.LoginForm;
+import com.landvibe.landlog.dto.LoginForm;
 import com.landvibe.landlog.domain.Member;
 import com.landvibe.landlog.repository.MemberRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 import static com.landvibe.landlog.ErrorMessage.*;
 
@@ -42,12 +40,12 @@ public class MemberService {
         return member;
     }
 
-    public Member login(LoginForm loginForm) throws Exception {
+    public Long login(LoginForm loginForm) {
         Member member = memberRepository.findByEmail(loginForm.getEmail())
-                .orElseThrow(() -> new Exception(NO_USER.message));
+                .orElseThrow(() -> new IllegalArgumentException(NO_USER.message));
         if (!member.getPassword().equals(loginForm.getPassword())) {
             throw new IllegalArgumentException(WRONG_PASSWORD.message);
         }
-        return member;
+        return member.getId();
     }
 }
